+++ Physics2D/Force/ParticleForceGenerator.cs
@@ -4,6 +4,6 @@ namespace Physics2D.Force
 {
     public abstract class ParticleForceGenerator
     {
-        public abstract void UpdateForce(Particle particle, float duration);
+        public abstract void UpdateForce(Particle particle, double duration);
     }
 }
