+++ Physics2D/Force/ParticleGravity.cs
@@ -11,7 +11,7 @@ namespace Physics2D.Force
             this.gravity = gravity;
         }
 
-        public override void UpdateForce(Object.Particle particle, float duration)
+        public override void UpdateForce(Object.Particle particle, double duration)
         {
             // 质量无限大的物体不受重力影响
             if (particle.InverseMass == 0f) return;
